refactor: syscall and terminal
